refactor: MovieViewHolder | bind 메서드 추가
